Rename vireoapi to match module name
