+++ source/io/module_eggShell.js
