+++ test-it/httptest1/loaderamdenvglobal.html
@@ -4,7 +4,7 @@
 The existing loaderglobal code should still be able to function -->
 <script data-main="loaderamdenvglobalmain" src="../../node_modules/requirejs/require.js"></script>
 <script src="../../dist/vireo.js"></script>
-<script src="../../source/io/module_vireoapi.js"></script>
+<script src="../../source/io/module_eggShell.js"></script>
 <script src="../../source/io/module_httpClient.js"></script>
 <script src="../../source/core/vireo.loader.js"></script>
 <script src="loaderglobal.js"></script>
