+++ test-it/httptest1/loaderglobal.html
@@ -1,7 +1,7 @@
 <html>
 <head>
 <script src="../../dist/vireo.js"></script>
-<script src="../../source/io/module_vireoapi.js"></script>
+<script src="../../source/io/module_eggShell.js"></script>
 <script src="../../source/io/module_httpClient.js"></script>
 <script src="../../source/core/vireo.loader.js"></script>
 <script src="loaderglobal.js"></script>
